Update Bush-2022.qmd
I added a couple of tools I think are important for the standardization of the protocol and added to the workflow not only the steps that an individual can do, but those taken by the centre to do a proper transition. Feel free to discuss and object about adding these!